Ok, I fixed LOTS of minor details that were causing warnings from the compilers. However, there is a major bit of broken code here: the CUSP library needs the NVCC compiler. The armadillo headers included by classes using CUSP cause NVCC to abort with an error.

+++ src/grids/cvt/alt_ellipsoid_cvt.cpp
@@ -615,7 +615,7 @@ void AltEllipsoidCVT::fillBoundaryPoints(int dim_num, int nb_nodes, int *seed, d
 
     // Now that we know how many boundary nodes we want out of the total
     // number of nodes in the domain, resize to that value
-    this->nb_bnd = nb_computed_bnd;
+    this->nb_bnd = (int)nb_computed_bnd;
     this->resizeBoundary(this->nb_bnd);
     //bndry_pts.resize(this->nb_bnd);
